Renamed package and added md file
