+++ java-shallow-deep-copy/src/main/java/com/baeldung/App.java
@@ -1,4 +1,4 @@
-package com.baeldung.java_shallow_deep_copy;
+package com.baeldung;
 
 /**
  * Hello world!
