+++ java-shallow-deep-copy/src/main/java/com/baeldung/data/BankAccountShallow.java
@@ -1,4 +1,4 @@
-package com.baeldung.java_shallow_deep_copy.data;
+package com.baeldung.data;
 
 public class BankAccountShallow extends BankAccount {
